Refactored: Datadog::Metrics --> Datadog::Core::Metrics
