+++ lib/datadog/core/metrics/metric.rb
@@ -1,0 +1,12 @@
+module Datadog
+  module Core
+    module Metrics
+      Metric = Struct.new(:type, :name, :value, :options) do
+        def initialize(*args)
+          super
+          self.options = options || {}
+        end
+      end
+    end
+  end
+end
