+++ spec/ddtrace/runtime/metrics_spec.rb
@@ -233,7 +233,7 @@ RSpec.describe Datadog::Runtime::Metrics do
 
       context 'when no services have been registered' do
         it do
-          is_expected.to include(*Datadog::Metrics.default_metric_options[:tags])
+          is_expected.to include(*Datadog::Core::Metrics::Client.default_metric_options[:tags])
           is_expected.to include('language:ruby')
         end
       end
@@ -244,7 +244,7 @@ RSpec.describe Datadog::Runtime::Metrics do
         before { services.each { |service| runtime_metrics.register_service(service) } }
 
         it do
-          is_expected.to include(*Datadog::Metrics.default_metric_options[:tags])
+          is_expected.to include(*Datadog::Core::Metrics::Client.default_metric_options[:tags])
           is_expected.to include('language:ruby')
           is_expected.to include(*services.collect { |service| "service:#{service}" })
         end
