+++ spec/ddtrace/transport/statistics_spec.rb
@@ -41,7 +41,7 @@ RSpec.describe Datadog::Transport::Statistics do
             expect(Datadog.health_metrics).to have_received(:send_metrics) do |metrics|
               expect(metrics).to be_a_kind_of(Array)
               expect(metrics).to have(1).item
-              expect(metrics).to include(kind_of(Datadog::Metrics::Metric))
+              expect(metrics).to include(kind_of(Datadog::Core::Metrics::Metric))
               expect(metrics.first).to have_attributes(
                 type: :api_responses,
                 name: nil,
@@ -70,7 +70,7 @@ RSpec.describe Datadog::Transport::Statistics do
             expect(Datadog.health_metrics).to have_received(:send_metrics) do |metrics|
               expect(metrics).to be_a_kind_of(Array)
               expect(metrics).to have(1).item
-              expect(metrics).to include(kind_of(Datadog::Metrics::Metric))
+              expect(metrics).to include(kind_of(Datadog::Core::Metrics::Metric))
               expect(metrics.first).to have_attributes(
                 type: :api_responses,
                 name: nil,
@@ -99,7 +99,7 @@ RSpec.describe Datadog::Transport::Statistics do
             expect(Datadog.health_metrics).to have_received(:send_metrics) do |metrics|
               expect(metrics).to be_a_kind_of(Array)
               expect(metrics).to have(1).item
-              expect(metrics).to include(kind_of(Datadog::Metrics::Metric))
+              expect(metrics).to include(kind_of(Datadog::Core::Metrics::Metric))
               expect(metrics.first).to have_attributes(
                 type: :api_responses,
                 name: nil,
@@ -128,7 +128,7 @@ RSpec.describe Datadog::Transport::Statistics do
             expect(Datadog.health_metrics).to have_received(:send_metrics) do |metrics|
               expect(metrics).to be_a_kind_of(Array)
               expect(metrics).to have(1).item
-              expect(metrics).to include(kind_of(Datadog::Metrics::Metric))
+              expect(metrics).to include(kind_of(Datadog::Core::Metrics::Metric))
               expect(metrics.first).to have_attributes(
                 type: :api_errors,
                 name: nil,
@@ -244,7 +244,7 @@ RSpec.describe Datadog::Transport::Statistics do
         expect(Datadog.health_metrics).to have_received(:send_metrics) do |metrics|
           expect(metrics).to be_a_kind_of(Array)
           expect(metrics).to have(1).item
-          expect(metrics).to include(kind_of(Datadog::Metrics::Metric))
+          expect(metrics).to include(kind_of(Datadog::Core::Metrics::Metric))
           expect(metrics.first).to have_attributes(
             type: :api_errors,
             name: nil,
