+++ spec/support/metric_helpers.rb
@@ -10,7 +10,7 @@ module MetricHelpers
     def metric_options(options = nil)
       return options unless options.nil? || options.is_a?(Hash)
 
-      Datadog::Metrics.metric_options(options)
+      Datadog::Core::Metrics::Client.metric_options(options)
     end
 
     def check_options!(options)
